have is_gzipped stand alone
