+++ photon_stream/io/JsonLinesReader.py
@@ -1,7 +1,7 @@
 import ujson as json
 import os
 import gzip
-
+from .is_gzipped_file import is_gzipped_file
 
 class JsonLinesReader:
     """
@@ -38,15 +38,3 @@ class JsonLinesReader:
         out = '{}('.format(self.__class__.__name__)
         out += self.path+')\n'
         return out
-
-
-def is_gzipped_file(path):
-    """
-    Check for gzip file, see https://tools.ietf.org/html/rfc1952#page-5
-
-    Reads in the first two bytes of a file and compares with the gzip magic 
-    numbers.
-    """
-    with open(path, 'rb') as f:
-        marker = f.read(2)
-        return marker[0] == 31 and marker[1] == 139
