+++ photon_stream/io/__init__.py
@@ -1,3 +1,4 @@
 from .JsonLinesReader import JsonLinesReader
 from . import binary
 from .jsonl2binary import jsonl2binary
+from .is_gzipped_file import is_gzipped_file
