+++ photon_stream/io/is_gzipped_file.py
@@ -1,0 +1,13 @@
+import gzip
+
+
+def is_gzipped_file(path):
+    """
+    Check for gzip file, see https://tools.ietf.org/html/rfc1952#page-5
+
+    Reads in the first two bytes of a file and compares with the gzip magic 
+    numbers.
+    """
+    with open(path, 'rb') as f:
+        marker = f.read(2)
+        return marker[0] == 31 and marker[1] == 139
